Partly fixed problem with non-ascii encoding in the TFS repository path
The problem is that this value is written to .git\config which is done via git config command. And it uses system default encoding to write values. So the support was extended to system default encoding which is somewhat fragile (git-tfs will be broken if system default encoding changed).

More general approach is to encode this value in base-64 before saving to git config. It will make config file unreadable but will be robust. Further, all legitimate TFS paths start with $ char which can't appear in base-64 encoded string, so we can use it. If we have pure ASCII value - write it as it is, otherwise, encode it into base-64. When reading value check first byte. If it is ASCII code of '$' - read value normally, otherwise decode it before use. I'll leave it for TODO list for now.