Fixing CloneTests.CloneWithtMixedUpCase().
This test came very close to detecting the bug in #174. However, the bug only manifests itself when the difference in casing occurs deeper than the folders that are the children of the repo folder (so "$/MyProject/Folder" and "$/MyProject/folder" worked, but "$/MyProject/Foo/Bar" and "$/MyProject/Foo/bar" didn't work). The fixed test passes when 3a1d996aadcca181412a72ccebde3056304eea5c is applied and fails without it.
Also, the line endings have been normalized to LF, which was the dominant line ending in the file (but seems to be contrary to most other files in the repo).