Download correct files for unshelve
Downloading of files was the responsibility of the items themselves,
but to optimize this, it was changed to requesting entire changesets
at once. However, this optimization broke downloading of shelvesets.

The changeset is now responsible for populating the workspace with the
correct files. This allows for different download strategies for real
changesets and shelvesets.