Handle parsing merge commit messages with multiple TFS changeset ids
The original code would null out currentCommit after parsing the FIRST changeset id in the
log message, and associate the null commit hash with any additional changeset ids in the message
for the same commit. This eventually leads to a null reference error when executing git commands
that require the hash.

The new code ignores all of the changeset ids in the commit message except for the last one, which
is the only one that matters.