Do not delete a directory that is already deleted
In some cases a directory that needs to be deleted was already deleted
because an ancestor directory was deleted. Consequently, TFS would
complain about it not existing in the workspace.

For example:
root/dir1/deleteme.txt
root/dir1/dir2/deleteme.txt

Because of the deletion of 'root/dir1/deleteme.txt', 'root/dir1' becomes a
candidate for deletion. It is deleted because it no longer contains any
items.
However, because of the deletion of 'root/dir1/dir2/deleteme.txt',
'root/dir1/dir2' also is a canditate for deletion. It also needs to be
deleted, but this is already taken care of with the deletion of its parent
directory.