Use ASCII.
This is just for simplicity. Version.cs doesn't need any non-ASCII
characters, and `git diff` works more reliably with ASCII than with
other encodings.

In general, Unicode is fine for files, and the .gitattributes branch
will help this. But for Version.cs, I'm making it ASCII, at least for
now.