Avoid resetting core.autocrlf when initializing branches.
As reported in issue #424 the core.autocrlf value was being reset to the
default when a tfs branch was initialized resulting in a broken branch when
autocrlf was set true. This patch passes the empty string to indicate the
setting should not be modified (as null is already being used to indicate
the default value should be used).

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>