Allow the work-item association regex to be overridden by user config.
This is intended to resolve issue #512 by allowing the user git config to
redefine the regular expression used to match work-item id's in the commit
message. This is useful in those cases where a project may be using an
alternative system to track bug id's.

The user may define the git config variable git-tfs.work-item-regex to
use an alternate regular expression. If this is undefined or en empty
string then the built-in default is used (matching # followed by digits).
If the expression fails to compile then no matching will be attempted.

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>