Limit version scope for querying merge information
When querying the merge information to determine the source changeset of a
merge, limit the version scope of the query to the bare minimum to improve
performance.

Since the QueryMerges overload taking strings for source and target calls
the overload with ItemSpecs, there is also no need to have a seperate post
VS2010 implementation. At least this is the case for v12.020827.3 of
Microsoft.TeamFoundation.VersionControl.Client.dll.