Added option to avoid replicating TFS merges during fetch.
Sometimes TFS guys do merge changesets apparently at random, while in git merge-commit has meaning that _all_ commits in the parent branches are merged. Thus preserving all of the merges sometimes do more harm than good. New setting is tfs.replicateTfsMerges and could be "0" or "1". This setting is turned on by default (i.e. if there's no such value in the git config file - it treats as "1") and is global (not tied to particular tfs remote/branch - mostly because on the per-tfs-remote basis it can do much confusion if configured differently in different places.