Remove confusing "cleaning" trace
In debug, when cleaning is done in the finally scope,
cleaning message is displayed BEFORE the error message
of the exception that occurs.

It's a little confusing because we think
that's the cleaning that went wrong while this
is a previous task that throw the exception...

So, remove the cleaning message.