Shorten workspace to mitigate ntfs path length limit
(See https://github.com/git-tfs/git-tfs/blob/master/doc/Set-custom-workspace.md
for problem description)

Now that the ".git\tfs\RemoteId\" directory is only used
to create a "workspace" directory and is cleaned after
each command run, we could instead use a unique and
common directory with a shorter path (".git\~w")!

This way we are less likely to hit the ntfs path length limit
because we are not dependent of the remote id. If we
succesfully clone the trunk, we will be able to fetch all the branches
(except if there is files deeper needless to say)