When a ref name is already used, find a strategy to give another one
An example of the case happening when cloning an existing tfs repository
was a branch created in "$/project/dev/my_branch" when in the past was
existing or still exists a branch with the path "$/project/dev"
(the first refs created prevent to create a folder with the same name).

The solution here is to add a `_` at the beginning of the expected ref name
until we found a ref name which is not used.