Fix fetching author after change to VSTS
Something changed on Visual Studio Team Services that meant the old
deprecated API we were using no longer returned the name/email we need,
resulting in commit authors being the raw user guid. This change
finishes off the change to the new API that was started previously and
gets it producing the display name and email address again.

* Read identity from GUID using new API.
* Update wrapper to use new TFS identity class.
* Map returned data based on what looked right when debugging.

Effectively reverts and finishes off the original change: "Revert
TeamFoundationIdentity because it changes the identity received" -
2b8eaecd993b1112b38da019d7664215b8ad8e49.

refs:
* https://social.msdn.microsoft.com/Forums/vstudio/en-US/c0fea096-92d5-4466-832c-43db0a4e65ae/how-to-retrive-user-alias-by-user-guid?forum=tfsgeneral
* https://msdn.microsoft.com/en-us/library/ff731923%28v=vs.120%29.aspx - IIdentityManagementService.ReadIdentities()

Fixes #934