Update TFS Client dependencies
Update the `Microsoft.TeamFoundationServer.*Client` references from `14.89.0` to `14.95.3`. The justification is that when we were on a TFS 2012 server, the current references worked. However, at the organization I work at, we recently migrated to TFS 2013. After migration, we got one of the following errors:

>ERROR] Policy: Internal error in Changeset Comments Policy. Error loading the Changeset Comments Policy polilture=neutral, PublicKeyToken=b03f5f7f11d50a3a' is not registered.). Installation instructions:
>No changes checked in.

and

>[ERROR] Policy: Internal error in Changeset Comments Policy. Error loading the Changeset Comments Policy policy (The policy assembly 'Microsoft.TeamFoundation.PowerTools.CheckinPolicies.ChangesetComments, Version=10.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a' is not registered.). Installation instructions:
>No changes checked in.

This fixes the issue reported as issue #959 by @JoachimC.