Build git-tfs as x86
When git-tfs is built as AnyCPU, it is unable to access checkin policy DLLs that are built as x86 when run on a x64 machine. Since checkin policy DLLs can only be x86 or AnyCPU (Visual Studio is x86 only), then making git-tfs x86 allows it to support all dlls that Visual Studio supports.