Fix broken unit tests; update for new LibGit2Sharp
When I branched off of master, there were 5 broken unit tests. The main
issue was that a branch named "branch" was created, and then a commit
would be made and a comparison of the resulting SHAs. However, what was
forgotten in these tests was to checkout the branch "branch" after it
was created prior to creating the next commit. These oversights were
corrected.

Also, with the new version of LibGit2Sharp, the unit tests needed
some updating to remove method calls to obsolete methods.

Also updated the ignore file to ignore NCrunch, Resharper, and
DevTracker files.

Signed-off-by: Craig E. Shea <crag.e.shea@gmail.com>