Make LoadFromVsFolder abstract
Because of the differences in installation folder layout between Visual
Studio 2010-2015 and the latest Visual Studo 2017, it is necessary to
allow each GitTfs Helper to define how to load an assembly from the
Visual Studio folder.

This was accomplished by creating a private `_LoadFromVsFolder` method
which calls the (now) protected and abstract `LoadFromVsFolder` method.
Each GitTfs Helper must now implement/override this method.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>