Stop using property `InitialChangeset` for branch information.
The global property `InitialChangeset` is used to store the first changeset the user want to clone from.

But this was also in some situations used to store the first changeset of a branch. This cause a problem as this property is global and will be applied to all branches. If `InitialChangeset` is set for a branch, the next cloned branch can not get the changesets before the changeset given this `InitialChangeset` if it is not updated for that branch to.

It is problematic to use a property for different information.