Remove GitTfs.Setup and the broken MSI file handling
Building the MSI is not working since 6732259 ("Moving source in
the 'src' folder", 2017-10-13) and when actually looking at the
GitHub releases, the MSI is not attached to any of them.

The MSI file was never distributed on any GitHub release and using
copy deployment zip file is much more conveniant, because it allows
having multiple versions available.

Therefore, remove the GitTfs.Setup code and the upload function
in the cake build file.