Introduce new paket dependency group VS2015
In prepration for introducing support for VS2017, introduce a new paket
group VS2015 where the specific Visual Studio 2015 dependencies are
collected. Later on, we are going to introduce a new group for VS2017, too.
So doing this minimal refactoring first makes sure we don't run into
any paket / group specific glitches.

The paket.lock file was updated manually (using an editor) instead of
running paket, which did change the versions of the referenced packages
which I wanted to avoid as I have no way to test VS2015 anymore.

For reviewing, to see that this is really a simple code move without any
other changes, it is best to review the changes to src/paket.lock with
  git diff --color-moved --anchored="    Moq (4.14.7)" --anchored="    structuremap (2.6.3)" -- src/paket.lock

For that purposes `paket install` was not run, as it makes further minor
adaptions to the paket.lock file. This will be done in a follow up commit.
The split was made help the reviewer by seperating the move and tool based
adaptions done on top.