Pin paket version to 5.251.0
Otherwise, the build is not completely reproducable and depending if there
is a new paket version available, due to its autoupgrade it will mess with
the checked in file src/.paket/Paket.Restore.targets and cause strange
conflicts.

This commit has proven usefull when dealing a long running branch of patches
to the paket.dependencies and/or paket.lock file.