Show note on how to fix problems with the checkin policy only once
In case multiple checkin policies are used and all of them fail to load,
the screen is flooded with multiple repetions of the hint/note for the user
on how to fix the error. Instead, show the note only once as the last item
in the output, which is easier to follow.