Add support for VS2019
This implements the support for Visual Studio 2019 by adding a VS2019
paket group. It is heavily based on the plugin for VS2017, as the
GitTfs.VS2017 plugin was initially copied to server as a base for this
new GitTfs.VS2019 plugin. The actual logic business logic is contained
in src\GitTfs.VsCommon\TfsHelper.Vs2017Base.cs and is shared between
both plugins.

HINT: To see what actually was changed after the GitTFS.VS2017 was copied,
use e.g. `git show -C50 -C HEAD` to spot the differences in the review.