Create skeleton TfsHelper.Vs2017Base.cs based on TfsHelper.Vs2012Base.cs
VS2017 and VS2019 need changes on how to interact with TFS and also need
a different mechanism for detection. Instead on deriving any future
TfsHelper from the existing TfsHelperVs2012Base used for VS2015, create
a new class acting as a common base for the VS2017 and VS2019 TfsHelper.
This has several benefits:
  - Not changing any code paths relevant for VS2015, as I do not have the
    setup to test this.
  - Easier to remove VS2015 in the future, if we no longer want to
    support it. It is quite old already today.

To make review easier, this commit simply copies TfsHelper.Vs2012Base.cs
to TfsHelper.Vs2017Base.cs, e.g.

  cp src/GitTfs.VsCommon/TfsHelper.Vs20{12,17}Base.cs

The file isn't used yet, e.g. not added to the build. Further changes
will be made on top, but doing it this way makes it absolutely clear what
changes are actually needed on top and therefore helps review.