Move method LoadFromVSFolder from class TfsHelperBase to its derived classes
Refactor the code by moving the LoadFromVSFolder method from TfsHelperBase
to its derived classed TfsHelperVs20{12,17}Base and make it abstract
in TfsHelperBase.

This gives us room to change the implementation in TfsHelperVs2017 but
still keeping the old logic for VS2015 intact. This is especially
helpfull as I can't test VS2015 here.

Furthermore, as now the last user of GetVsInstallDir method in the base
class is removed, remove the abstract method in the base class and
the override keyword from the prior implementations.

This is a preparational step because we want to change semantics about the
return value of GetVsInstallDir for VS2017 support, but returning
different semantic values dependening on the derived class would violate
the liskov substitution principle. With this change, this is no longer
the case as there is no common base class which has this method
implemented and both derived classes can implment the method however
they see fit.