Add support for checkin policies for VS2017 and VS2019
Visual Studio versions >= 2017 (or is it actually 2015 ?) store the
necessary information for e.g. CheckinPolicies no longer in the registry.
Instead, they use a private registry usually found in

  C:\Users\<USER>\AppData\Local\Microsoft\VisualStudio\15.0_xxxxxx\privateregistry.bin

where the xxxxxx value is differing between the different installed Visual
Studio editions and/or minor versions.

This commit introduces the ExternalSettingsManager which loads this private registry
and we can then add the VS extension folders gathered from the SettingsStore to
our assembly search path.

Sadly, this only works for 32bit process and crashes in a 64bit process as soon
as a method tries to create an ExternalSettingsManager instance. Therefore
this is put behind a check so we only do it in a 32bit process.

This implies that the checkin policies are only working in 32bit mode!

In addition to the search patsh gathered from the ExternalSettingsManager, add
  <vsInstallDir>\Common7\IDE\CommonExtensions\Microsoft\TeamFoundation\Team Explorer
to the search path, as otherwise some of the checkin policies used in my dev
environment here didn't load because a library in that folder wasn't
found. This indicated to me that we still are not 100% compatibile to
what Visual Studio does, but at least this fixes all the custom checkin
policies here, so it can't be that bad.