Add support for VS2017
This implements the support for VS2017 by implementing the necessary
logic to enumerate the installed VS intances. As of VS2017, multiple
Visual Studio versions can be installed in parallel, e.g. a Enterprise
and a Professional edition. Furthermore, it also supports installing
the same edition in multiple different versions.

The logic implemented in this commit will always use the first version found
with a matching major. As it isn't clear that the used underlying mechanism
to iterate over the installed versions is deterministic, this means that
if multiple versions of VS2017 are installed in parallel, we may find
different instances between seperate calls.

A typical user doesn't have multiple versions of the same VS major
installed, but instead it is much more common to have different major
vesions like e.g. VS2017 and VS2019 installed.

Therefore, finding the first VS2017 version is acceptable for now. Selecting
different editions or different versions of the same major can be done
later if it is needed at all, especially as testing this requires
a lot of work due to the need of installing multiple versions or editions
of Visual Studio.