TfsHelperVS2017Base: override GetDialogAssemblyPath and make it abstract in base class
For VS2017 and VS2019, the dialog assembly
  Microsoft.TeamFoundation.VersionControl.ControlAdapter
specified in the base class is actually in
  <vsInstallDir>\Common7\IDE\CommonExtensions\Microsoft\TeamFoundation\Team Explorer
and not in
  <vsInstallDir>\Common7\IDE\PrivateAssemblies

Override GetDialogAssemblyPath() to specify the correct location.

As it is already overriden in TfsHelper.Vs2015 which is the only other
class currently implementing TestHelperBase, the common shared
implemention in the base class is not used at all.

Therefore, make it abstract in the base class.