Adapt TfsHelperVS2017Base.LoadFromVsFolder method to changed semantics
The method GetVsInstallDir returns the path to the toplevel folder
of the found Visual Studio installation, e.g.

  C:\Program Files (x86)\Microsoft Visual Studio\2017\Enterprise

which is semantically different to what it returns for VS2015 support.
At least from reading the code, it looks like it returnes a path like

  C:\Program Files (x86)\Microsoft Visual Studio\2015\Common7\IDE

there. To handle that semantic difference, make the path adaptions and
while at it, also make a small refactoring to make it easier to support
loading assemblies from multiple search paths. Currently, in this commit
no additional search paths are added, only the necessary adaptions due to
the different semantic return paths of GetVsInstallDir for VS2017 are made.