Make the ITfsHelper a singleton so it isn't re-created for every TFS branch
Prior to this commit, the concrete TfsHelper is created afresh, especially
for each and every TFS branch it is created again. This makes it wasteful
if there are any expensive operations running in the concrete TfsHelper,
like it is done for the VS2017 TfsHelper which wants to detect the Visual
Studio installation directories only once and not multiple times.

Therefore, mark the concrete TfsHelper as a singleton to ensure that it
is created only once. This gets rid of the excessive tracing when detecting
the VS versions.