Remove the check if the required Visual Studio version is installed
... before trying load the corresponding plugin. With newer Visual Studio
versions, the check doesn't work anyhow, as the corresponding entries
in the registry are no longer written by those VS versions.

Therefore, remove the optional parameter `isVisualStudioRequired` to
`TryLoadVsPluginVersion` and get rid of the code reading from the registry,
as it is no longer used anymore.

While at it, make the code trying to load the TfsPlugin more generic so
that we don't need to change multiple places when we want to support a
new major Visual Studio version.
Furthermore, rename the SupportedVersion to SupportedVersions (plural), as
we are actually dealing with multiple versions.