Show note in case of checkin policy problems during shelve
Checkin Policies only work 32bit mode and for checkin and rcheckin an
appropriate note on how to fix the problem is shown to the user if there
are problems detected when evaluating the checkin policies.

Show the same note in `git-tfs shelve` when the user explicitely
requested via the commandline parameter `-p` to evaluate the checkin
policies.