Extend documentation on how to troubleshot checkin policies
Add seperate sections on checkin policies for VS2017 and VS2019 and
explain how loading a registry hive works so it might get easier for someone
to futher debug into failures for the checkin policies.

Furthermore, add a section that checkin policies for VS2015 are not
supported.

The old content is left intact, as it contains information which might
still be valuable, but it is marked as obsolete so it is clear that
it is only for older git-fs versions.