ProcessCheckinNoteCommand: don't trim the commit message unconditionally
Prior to this commit, trailing whitespace (including CRLF or LF) was trimmed
from the message, **even** if no matching note was found. This was at
least unexpected. If such an unconditional cleanup should be done, it
better should be a step of its own.

Change it so that the whitespace trimming is only done when an actual note
was found (and therefore the message modified). This copies the
behaviour of other steps which make modifications to the commit message.

Removing this altoegether can be considered a behaviour change.
In practice, no one will notice as `git commit` itself will remove
trailing whitespace with `--cleanup=strip`, which is the default.
So in practice, there shouldn't be any trailing empty lines in the git
commits.