.editorconfig: don't force end_of_line to CRLF
We are normalizing the line endings via .gitattributes by setting the
text=auto attribute on all files. This ensures that git will check in
any file it detects as a text file with LF line ending and on checkout
(if no core.eol is explictely configured) it will use the native line
ending for the platform.

Meaning that for people who either edit the code in Linux or have
  core.eol = lf
in their git configuration, hardcoding the line ending causes endless
pain as now any editor using the .editorconfig will change the line
ending on these systems.

Therefore, comment it out and simply rely on git to convert the line
endings appropriately.