Remove support for TFS 2008
The check `Is2008OrOlder` calls _server.ConfigurationServer and checks
if the returned configurationsever is null. This call causes an error
in case a PAT (Personal Access Token) is used for authentication.
(Code for PAT is not yet submitted).

Remove the old support for 2008 as it is very unlikely that anyone is still
running a 12 year old TFS server. As a ripple effect, the check for
`CanGetBranchInformation` became also useless, therefore remove the method
and adapt its callers, too. In addition, two now useless tests were removed.