Add reproducer for #1398
git-tfs explicitely sets `core.autocrlf = false` during `git tfs clone` to
ensure that we don't do line ending conversion. In #1398 it was reported
that when cloning with a gitignore file, e.g.

  git-tfs clone <TFS-SERVER> <TFS-REPO> --gitignore=<PATH_TO-GITIGNORE>

then the cloned repo still has its line endings converted, assuming that the
user has core.autocrlf = true in his system or global config, which is the
default on windows.

Add a reproducer testcase showing this behaviour which currently fails.