Convert CreateRemote(RemoteInfo, String.Empty, null) to one parameter version
Due to the logic in the CreateRemote method, giving a String.Empty for
the `autocrlf` parameter or null for the `ignorecase` paramter will do
nothing - or in other words behaves the same as calling the other
overload with only the `RemoteInfo` parameter.

Simplify the code by replacing these calls with the one parameter
version of the `CreateRemote(RemoteInfo)` method. This is a straight
forward refactoring without changing any semantics.