Set core.autocrlf and core.ignorecase in DoGitInitDb to fix #1398
Both of those variables are global (= meaning not per Tfs remote) so set
them earlier when the repository is created, before any other access
to the libgit2 object database.

This fixes the reproducer for #1398 added in b272770873 as test
LineNotNormalizedWhenGitIgnoreGivenAndAutocrlfFalse_Issue1398

The underlying issue seems to be a problem in libgit2 (or libgit2sharp)
as it doesn't re-evaluate the config setting in core.autocrlf after a
commit was made, meaning it didn't use the explicit core.autocrlf=false
value we set when commiting the .gitignore file.

By setting the variable earlier in the init process, we work around this
problem and as this was the only remaining caller of the 3 parameter
`IGitRepository.CreateTfsRemote` overload, remove it.

This commit changes the behaviour of the .gitignore file checked in, meaning
before this commit, the behaviour was like this:
  IF    a. the user as core.autocrlf set to true in system or global config
  AND   b. the given --gitignore=<GITIGNORE_FILE> file has CRLF line endings
  AND   c. git-tfs init was called --autocrlf=false or no --autocrlf
           parameter at all

  THEN  the checked in .gitignore will has its line endings converted to LF
        in the repo.

With this commit, the line ending conversion is consistently done,
either for all the files (including the .gitignore) or for none at all,
so this minor behaviour change is actually an improvement.

Fixes #1398