IGitTfsRemote: remove unused properties UseGitIgnore and GitIgnorePath
In 1a60ec47 ("InitOptions: Promote --gitignore to RemoteOptions",
2018-08-10) the property GitIgnorePath was introduced, but no user outside
of the class was ever introduced needing that property.

The same applies to UseGitIgnore, which was introduced in 8eca32f4
("GitTfsRemote: Cleanup gitignore support configuration handling", 2018-08-11)
and only exposed at the class but not at the interface IGitTfsRemote.

Remove both of them as they currently don't have a purpose.