Remove unneccessary abstraction IItemDownloadStrategy by inlining its code
The DownloadFile method had in the past multiple different implementations,
using a stream or using a real file depending on the VS plugin version.
It was therefore implemented as a strategy object in the concrete TfsPlugin.

As of today, there is only one strategy used, so the interface and the
implementation doesn't serve any purpose anymore. Inline the code and
remove the IItemDownloadStrategy interface and its one and only
implementation.