Fix crash with checkin policy support and no installed extensions
If checkin policies are enabled via enable_checkin_policies_support.bat
the git-tfs binary is modified via CorFlags to run as a 32bit process,
as the checkin policies are only available in 32bit (not sure if this
is still the case for VS2022, but at least for VS2019 and VS2017 this is
the case).

The code uses internally a different code path if it runs as 32bit process,
as the corresponding libraries to enable loading the private registry
only works in 32bit. Commit a515071e ("Add support
for checkin policies for VS2017 and VS2019", 2020-11-02) introduced
the support for loading the private registry where e.g. installed
extensions and/or code review policies are stored, but it accidentally
had some debugging leftovers trying to access a specific registry key.

This key does not exist if no VS extensions and/or checkin policies are
installed and therefore made the git-tfs application terminate with an
exception.

Remove this dead code and fix this error, which was noticed when testing
the VS2022 support whitout any extensions installed.