Remove setter for TfsChangeset.Summary
The setter of the Summary property is not exposed on the ITfsChangeset
interface level and during the lifetime of the TfsChangeset object, the
property Summary should not change. Therefore prefere to initalize it
once in the ctor by adding it as another parameter to make this
immutability obvious by removing the setter.