Always favor the author from the authors file
In case the author comes from the TFS identity it was not replaced by a match in the authors file.
The authors file should always win because the identity on the TFS may be outdated. If the author should not be replaced simply do not include the mapping in the authors file.