Make fetch abort at any time by special temp file
If the "up to changeset" value turns out to be too high I want to be able to abort the fetch at any time without just killing the process. Placing a file in the temp directory seems like a simple solution to me.
Also inform the user about the file. In case it is forgotten to be deleted.