Fix handling of renamed branches for clone/fetch (#1493)
* In GitTfsRemote.FetchWithMerge(), apply special handling to initial
branch commits that are renaming, as rename destinations will have
these.
In GitTfsRemote.InitTfsRemoteOfChangeset(), when iterating over branch
ancestry, call InitBranch() on the ancestor branch, not the target
branch.  Also only request fetching of parents for the most distant
ancestor.  Also when handling renamed branches, assign the renameResult
for successors.
In GitTfsRemote.InitTfsBranch(), pull the initial commit hash out of
renameResult, if provided.

---------

Co-authored-by: Philippe Miossec <pmiossec@gmail.com>